requirements.txt
amqp==2.5.1
anyjson==0.3.3
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-allauth==0.39.1
django-appconf==1.0.3
django-celery==3.3.1
django-celery-email==2.0.2
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.22
kombu==4.6.4
more-itertools==7.2.0
oauthlib==3.1.0
pep8==1.7.1
Pillow==6.1.0
pkg-resources==0.0.0
psycopg2==2.8.3
PyJWT==1.7.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
whitenoise==4.1.3
zipp==0.6.0
